Move the random_bsr_matrix to sparse.utils
