+++ python/tvm/topi/nn/sparse.py
@@ -470,38 +470,6 @@ def try_get_sparse_input(args):
     return sparse_input_map
 
 
-def random_bsr_matrix(m, n, bs_r, bs_c, density, dtype):
-    """Generate a random sparse matrix in bsr format.
-
-    Returns
-    -------
-    scipy.sparse.bsr_matrix
-    """
-    # pylint: disable=import-outside-toplevel
-    import numpy as np
-    import itertools
-    import scipy.sparse as sp
-
-    y = np.zeros((m, n), dtype=dtype)
-    assert m % bs_r == 0
-    assert n % bs_c == 0
-    nnz = int(density * m * n)
-    num_blocks = int(nnz / (bs_r * bs_c)) + 1
-    candidate_blocks = np.asarray(list(itertools.product(range(0, m, bs_r), range(0, n, bs_c))))
-    assert candidate_blocks.shape[0] == m // bs_r * n // bs_c
-    chosen_blocks = candidate_blocks[
-        np.random.choice(candidate_blocks.shape[0], size=num_blocks, replace=False)
-    ]
-    # pylint: disable=invalid-name
-    for (r, c) in chosen_blocks:
-        y[r : r + bs_r, c : c + bs_c] = np.random.randn(bs_r, bs_c)
-    s = sp.bsr_matrix(y, blocksize=(bs_r, bs_c))
-    assert s.data.shape == (num_blocks, bs_r, bs_c)
-    assert s.indices.shape == (num_blocks,)
-    assert s.indptr.shape == (m // bs_r + 1,)
-    return s
-
-
 def sparse_add(dense_data, sparse_data, sparse_indices, sparse_indptr):
     """
     Computes sparse-dense addition
