+++ tutorials/auto_scheduler/tune_network_x86.py
@@ -50,7 +50,7 @@ import numpy as np
 import tvm
 from tvm import relay, auto_scheduler
 from tvm.relay import data_dep_optimization as ddo
-from tvm.topi.nn.sparse import random_bsr_matrix
+from tvm.topi.sparse.utils import random_bsr_matrix
 import tvm.relay.testing
 from tvm.contrib import graph_runtime
